Deploy app via Github Actions (#3)
Intention is to delegate the deployment process to Github Actions so that developers do not necessarily have to set up Google Cloud SDK. Unit tests will automatically run for all PRs that target the 'master' branch and any pushes to the 'master branch. Additionally, on pushing to the 'master' branch, the app will automatically be deployed to the testing environment. A manual deployment workflow is available to deploy to any environment in an ad-hoc fashion.